Move VelocityAdvection into solve_nonhydro.py
SolveNonhydro is the only user of VelocityAdvection: it constructs it
once and calls its two run methods from its own predictor and corrector
steps. The class moves verbatim (244 lines, byte-identical) and
velocity_advection.py goes. The test now constructs the class from
solve_nonhydro, and the docs reference follows it.
