+++ model/atmosphere/dycore/docs/dycore_numerics_advection.rst
@@ -3,5 +3,5 @@ Dycore numerical documentation (velocity advection)
 
 This page documents the advection part of the dycore.
 
-.. autoscidoc:: icon4py.model.atmosphere.dycore.velocity_advection.VelocityAdvection.run_predictor_step
+.. autoscidoc:: icon4py.model.atmosphere.dycore.solve_nonhydro.VelocityAdvection.run_predictor_step
    :no-index:
